// lib/api/types.gen.ts
export type IncidentEditModel = {
    startDate: string;
    stopDate: string;
    aCrew: number;
    bCrew: number;
    cCrew: number;
    dCrew: number;
    dayCrew: number;
};
// ...
export type StarsByCrewModel = {
    crew: string;
    starCount: number;
};